+++ tests/testthat/test-allele_tests.R
@@ -1,3 +1,3 @@
 test_that("AlleleFreq works", {
-  expect_is(allele.freq(hetTgen(readVCF(paste0(path.package("rCNV"), "/example.raw.vcf.gz")),"GT"),"ind"), class="data.frame")
+  expect_type(allele.freq(hetTgen(readVCF(vcf.file.path=paste0(path.package("rCNV"), "/example.raw.vcf.gz")),"GT"),"ind"), "data.frame")
 })