+++ tests/testthat/test-cpm.R
@@ -1,3 +1,3 @@
 test_that("Normalization works", {
-  expect_is(cpm.normal(ADtable), class="list")
+  expect_type(cpm.normal(ADtable), "list")
 })