+++ tests/testthat/test-depthVsSample.R
@@ -1,3 +1,3 @@
 test_that("Sims works", {
-  expect_is(depthVsSample(cov.len=50,sam.len=100), class="matrix")
+  expect_type(depthVsSample(cov.len=50,sam.len=100), "double")
 })