+++ tests/testthat/test-dupGet.R
@@ -1,3 +1,3 @@
 test_that("detection works", {
-  expect_is(dupGet(alleleINF), class="data.frame")
+  expect_type(dupGet(alleleINF), "list")
 })