+++ tests/testthat/test-maf.R
@@ -1,3 +1,3 @@
 test_that("MAF works", {
-  expect_is(maf(ADtable), class="data.frame")
+  expect_type(maf(ADtable), "list")
 })