+++ tests/testthat/test-readVCF.R
@@ -1,3 +1,3 @@
 test_that("readVCF works", {
-  expect_is(readVCF(paste0(path.package("rCNV"), "/example.raw.vcf.gz")), class="list")
+  expect_type(readVCF(vcf.file.path=paste0(path.package("rCNV"), "/example.raw.vcf.gz")), "list")
 })