+++ tests/testthat/test-sim.R
@@ -1,4 +1,4 @@
 test_that("simulation works", {
-  expect_is(sim.als(n=50,nrun=100), class="list")
+  expect_type(sim.als(n=50,nrun=100), "list")
 })