+++ tests/testthat/test-vcf.R
@@ -1,3 +1,3 @@
 test_that("stats works", {
-  expect_is(vcf.stat(readVCF(paste0(path.package("rCNV"), "/example.raw.vcf.gz"))), class="data.frame")
+  expect_type(vcf.stat(readVCF(vcf.file.path=paste0(path.package("rCNV"), "/example.raw.vcf.gz"))), "data.frame")
 })